# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inquira",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/paths": {
      "get": {
        "tags": [
          "V1 Runtime"
        ],
        "summary": "Get Workspace Paths",
        "operationId": "get_workspace_paths_api_v1_workspaces__workspace_id__paths_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacePathsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/execute": {
      "post": {
        "tags": [
          "V1 Runtime"
        ],
        "summary": "Execute Workspace Code",
        "operationId": "execute_workspace_code_api_v1_workspaces__workspace_id__execute_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/datasets/{table_name}/schema": {
      "get": {
        "tags": [
          "V1 Runtime"
        ],
        "summary": "Get Workspace Dataset Schema",
        "operationId": "get_workspace_dataset_schema_api_v1_workspaces__workspace_id__datasets__table_name__schema_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "V1 Runtime"
        ],
        "summary": "Save Workspace Dataset Schema",
        "operationId": "save_workspace_dataset_schema_api_v1_workspaces__workspace_id__datasets__table_name__schema_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSchemaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/datasets/{table_name}/preview": {
      "get": {
        "tags": [
          "V1 Runtime"
        ],
        "summary": "Get Workspace Dataset Preview",
        "operationId": "get_workspace_dataset_preview_api_v1_workspaces__workspace_id__datasets__table_name__preview_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          },
          {
            "name": "sample_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(random|first)$",
              "default": "random",
              "title": "Sample Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/schemas": {
      "get": {
        "tags": [
          "V1 Runtime"
        ],
        "summary": "List Workspace Schemas",
        "operationId": "list_workspace_schemas_api_v1_workspaces__workspace_id__schemas_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetPreviewResponse": {
        "properties": {
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "sample_type": {
            "type": "string",
            "title": "Sample Type"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "table_name",
          "sample_type",
          "row_count",
          "data"
        ],
        "title": "DatasetPreviewResponse"
      },
...
      "DatasetSchemaResponse": {
        "properties": {
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "context": {
            "type": "string",
            "title": "Context",
            "default": ""
          },
          "columns": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "table_name",
          "columns"
        ],
        "title": "DatasetSchemaResponse"
      },
      "ExecuteRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Python code to execute"
          },
          "timeout": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Timeout",
            "description": "Max execution time in seconds",
            "default": 60
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ExecuteRequest"
      },
      "ExecuteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "stdout": {
            "type": "string",
            "title": "Stdout",
            "default": ""
          },
          "stderr": {
            "type": "string",
            "title": "Stderr",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "result_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Type"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ExecuteResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveSchemaRequest": {
        "properties": {
          "context": {
            "type": "string",
            "title": "Context",
            "default": ""
          },
          "columns": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "columns"
        ],
        "title": "SaveSchemaRequest"
      },
      "SchemaListResponse": {
        "properties": {
          "schemas": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Schemas"
          }
        },
        "type": "object",
        "required": [
          "schemas"
        ],
        "title": "SchemaListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspacePathsResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "workspace_dir": {
            "type": "string",
            "title": "Workspace Dir"
          },
          "duckdb_path": {
            "type": "string",
            "title": "Duckdb Path"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "workspace_dir",
          "duckdb_path"
        ],
        "title": "WorkspacePathsResponse"
      },
...
    }
  }
}